+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Added
 
 - Add worker to get message, download PDF from S3, extract text and proccess in LLM.
+- Implement MongoDB and improve schema to invoice.